+++ src/main/java/ru/shipownerproject/databases/vesselrepository/vesseltyperepository/VesselTypeRepository.java
@@ -3,5 +3,5 @@ package ru.shipownerproject.databases.vesselrepository.vesseltyperepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.shipownerproject.models.vessels.type.VesselType;
 
-public interface VesselTypeRepository extends JpaRepository<VesselType, Short> {
+public interface VesselTypeRepository extends JpaRepository<VesselType, Byte> {
 }
